Put all commands into one file
